Start python implementation to see how it works out.
