+++ samsa/client.py
@@ -19,12 +19,18 @@ import itertools
 import json
 import logging
 
-from samsa import handlers
+from samsa import handlers, pysamsa
 from samsa.common import Broker, Topic, Partition
 from samsa.connection import BrokerConnection
 from samsa.exceptions import ImproperlyConfiguredError
 from samsa.protocol import MetadataRequest, MetadataResponse
 from zlib import crc32
+
+try:
+    import rd_kafka
+    RD_KAFKA_PRESENT = True
+else:
+    RD_KAFKA_PRESENT = False
 
 
 logger = logging.getLogger(__name__)
@@ -39,118 +45,37 @@ class SamsaClient(object):
     :ivar brokers: The :class:`samsa.common.Broker` map for this cluster.
     :ivar topics: The :class:`samsa.common.Topic` map for this cluster.
     """
-    def __init__(self, hosts='127.0.0.1:9092', handler=None, timeout=30):
-        """Create a connection to a Kafka cluster
+    def __init__(self,
+                 hosts='127.0.0.1:9092',
+                 use_greenlets=False,
+                 timeout=30,
+                 use_rdkafka=True):
+        """Create a connection to a Kafka cluster.
 
-        :param handler: Async handler.
-        :type handler: :class:`samsa.handlers.Handler`
+        :param hosts: Comma separated list of seed hosts to used to connect.
+        :param use_greenlets: If True, gevent will be used instead of threading.
+        :param timeout: Connection timeout, in seconds.
+        :param use_librdkafka: Use rd_kafka, if installed.
         """
         self._seed_hosts = hosts
         self._timeout = timeout
         self.brokers = {}
-        self.handler = handler or handlers.ThreadingHandler()
+        self.handler = None if use_greenlets else handlers.ThreadingHandler()
         self.topics = {}
-        self.update_cluster()
-
-    def _get_metadata(self):
-        """Get fresh cluster metadata from a broker"""
-        if self.brokers:
-            brokers = self.brokers.values()
+        self.use_rdkafka = RD_KAFKA_PRESENT and use_rdkafka
+        if self.use_rdkafka:
+            logger.info('Using rd_kafka extensions.')
+            raise NotImplementedError('Not yet')
         else:
-            brokers = [Broker(-1, host, port, self.handler, self._timeout)
-                       for host, port
-                       in [h.split(':') for h in self._seed_hosts.split(',')]]
-        for broker in brokers:
-            try:
-                return broker.request_metadata()
-            except Exception, e:
-                logger.warning('Unable to connect to broker %s:%s',
-                               broker.host, broker.port)
-                raise
-        raise Exception('Unable to connect to a broker to fetch metadata.')
-
-    def _update_brokers(self, broker_metadata):
-        # Remove old brokers
-        removed = set(self.brokers.keys()) - set(broker_metadata.keys())
-        for id_ in removed:
-            logger.info('Removing broker %s', self.brokers[id_])
-            self.brokers.pop(id_)
-        # Add/update current brokers
-        for id_,meta in broker_metadata.iteritems():
-            if id_ not in self.brokers:
-                self.brokers[id_] = Broker(meta.id, meta.host, meta.port,
-                                           self.handler, self._timeout)
-                logger.info('Adding new broker %s', self.brokers[id_])
-            else:
-                broker = self.brokers[id_]
-                if meta.host == broker.host and meta.port == broker.port:
-                    continue # no changes
-                if broker.connected:
-                    broker.disconnect()
-                logger.info('Updating broker %s', broker)
-                broker.host = meta.host
-                broker.port = meta.port
-                logger.info('Updated broker to %s', broker)
-
-    def _update_topics(self, topic_metadata):
-        # Remove old topics
-        removed = set(self.topics.keys()) - set(topic_metadata.keys())
-        for name in removed:
-            logger.info('Removing topic %s', self.topic[name])
-            self.topics.pop(name)
-        # Add/update partition information
-        for name,meta in topic_metadata.iteritems():
-            if name not in self.topics:
-                self.topics[name] = Topic(name)
-                logger.info('Adding topic %s', self.topics[name])
-            # Partitions always need to be updated, even on add
-            self._update_partitions(self.topics[name], meta.partitions)
-
-    def _update_partitions(self, topic, partition_metadata):
-        # Remove old partitions
-        removed = set(topic.partitions.keys()) - set(partition_metadata.keys())
-        for id_ in removed:
-            logger.info('Removing partiton %s', topic.partitons[id_])
-            self.brokers.pop(id_)
-        # Make sure any brokers referenced are known
-        all_brokers = itertools.chain.from_iterable(
-            [meta.leader,]+meta.isr+meta.replicas
-            for meta in partition_metadata.itervalues()
-        )
-        if any(b not in self.brokers for b in all_brokers):
-            raise Exception('TODO: Type this exception')
-        # Add/update current partitions
-        for id_,meta in partition_metadata.iteritems():
-            if meta.id not in topic.partitions:
-                topic.partitions[meta.id] = Partition(
-                    topic, meta.id, self.brokers[meta.leader],
-                    [self.brokers[b] for b in meta.replicas],
-                    [self.brokers[b] for b in meta.isr]
-                )
-                logger.info('Adding partition %s', topic.partitions[meta.id])
-            partition = topic.partitions[id_]
-            # Check leader
-            if meta.leader != partition.leader.id:
-                logger.info('Updating leader for %s', partition)
-                partition.leader = self.brokers[meta.leader]
-            # Check replica and In-Sync-Replicas lists
-            if sorted(r.id for r in partition.replicas) != sorted(meta.replicas):
-                logger.info('Updating replicas list for %s', partition)
-                partition.replicas = [self.brokers[b] for b in meta.replicas]
-            if sorted(i.id for i in partition.isr) != sorted(meta.isr):
-                logger.info('Updating in sync replicas list for %s', partition)
-                partition.isr = [self.brokers[b] for b in meta.isr]
-
+            self.cluster = pysamsa.Cluster(self._seed_hosts, self.handler)
 
     def update_cluster(self):
-        """Update known brokers and topics
+        """Update known brokers and topics.
 
-        We actually want to update this all in place, rather than paving over
-        previously known topics/partitions with new information. We do this
-        so that a topic can call ``self.client.update()`` to find new partition
-        leaders. If we replaced ``self.topics`` with a new dict, then that
-        partition info wouldn't be updated and that topic would be orphaned.
+        Updates each Topic and Broker, adding new ones as found,
+        with current metadata from the cluster.
         """
-        metadata = self._get_metadata()
-        self._update_brokers(metadata.brokers)
-        self._update_topics(metadata.topics)
+        # TODO: This is *so* not thread-safe, but should be rare.
+        #       Consider making a single lock while this runs to basically
+        #       stop the driver.
+        self.cluster.update()
