+++ samsa/pysamsa/__init__.py
@@ -1,0 +1,1 @@
+from cluster import Cluster
